# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Flashcards help you learn more vocabulary anytime, anywhere.",
        "title": "Word Flashcard API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/dictionary/{word}": {
            "get": {
                "description": "Get dictionary definition and pronunciation for a given word from Cambridge Dictionary API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dictionary"
                ],
                "summary": "Search dictionary for word definition",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Word to search for",
                        "name": "word",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dictionary definition found successfully",
                        "schema": {
                            "$ref": "#/definitions/word-flashcard_internal_models.DictionaryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - Missing word parameter",
                        "schema": {
                            "$ref": "#/definitions/word-flashcard_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found - Word not found in the dictionary",
                        "schema": {
                            "$ref": "#/definitions/word-flashcard_internal_models.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad gateway - Dictionary service unavailable",
                        "schema": {
                            "$ref": "#/definitions/word-flashcard_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "word-flashcard_internal_models.DefinitionInfo": {
            "type": "object",
            "properties": {
                "definition": {
                    "type": "string"
                },
                "example": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "word-flashcard_internal_models.DictionaryResponse": {
            "type": "object",
            "properties": {
                "meanings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/word-flashcard_internal_models.MeaningInfo"
                    }
                },
                "phonetics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/word-flashcard_internal_models.PhoneticInfo"
                    }
                }
            }
        },
        "word-flashcard_internal_models.ErrorCode": {
            "type": "string",
            "enum": [
                "invalid_request",
                "validation_error",
                "not_found",
                "conflict",
                "internal_error",
                "upstream_unavailable"
            ],
            "x-enum-varnames": [
                "ErrCodeInvalidRequest",
                "ErrCodeValidationError",
                "ErrCodeNotFound",
                "ErrCodeConflict",
                "ErrCodeInternalError",
                "ErrCodeUpstreamUnavailable"
            ]
        },
        "word-flashcard_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/word-flashcard_internal_models.ErrorCode"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "word-flashcard_internal_models.MeaningInfo": {
            "type": "object",
            "properties": {
                "definitions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/word-flashcard_internal_models.DefinitionInfo"
                    }
                },
                "partOfSpeech": {
                    "type": "string"
                }
            }
        },
...
        "word-flashcard_internal_models.PhoneticInfo": {
            "type": "object",
            "properties": {
                "audio": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                }
            }
        },
...
    }
}